refactor(task): move the fixtures
