+++ tasks/tests/test_model.py
@@ -1,43 +1,14 @@
-# Standard Library
-import datetime
-
 # Django
 from django.contrib.auth.models import User
 from django.urls import reverse
-from django.utils import timezone
 
 # Third Party
 import pytest
-from model_bakery import baker
 from rest_framework import status
 from rest_framework.test import APIClient
 
 # Locals
 from ..models import Task
-
-
-@pytest.fixture
-def create_task(user):
-    def _create_task():
-        return baker.make(Task, owner=user)
-
-    return _create_task
-
-
-@pytest.fixture
-def create_done_task(create_task):
-    def _create_done_task():
-        task = create_task()
-        task.done()
-        task.save()
-        return task
-
-    return _create_done_task
-
-
-@pytest.fixture
-def tomorrow():
-    return timezone.make_aware(datetime.datetime.now() + datetime.timedelta(days=1))
 
 
 @pytest.mark.django_db
